[checkHelper] fix

+++ CheckHelper.cs
@@ -19,19 +19,31 @@ namespace DevBlackListSoamChecker
                 {
                     if (admin.user.id != TgApi().getDefaultApiConnection().getMe().id)
                     {
-                        var result = TgApi.getDefaultApiConnection().sendMessage(
-                            Temp.ReportGroupID,
-                            "[加群測試(不用理會此訊息)](tg://user?id=" + admin.user.id.ToString() + ")",
-                            ParseMode: TgApi.PARSEMODE_MARKDOWN);
+                        var result = TgApi.getDefaultApiConnection().getChatMember(Temp.ReportGroupID , admin.user.id);
                         if (result.ok)
-                        {
-                            TgApi.deleteMessage(Temp.ReportGroupID, result.message_id);
-                            status = true;
-                            break;
-                        }
+                            if(result.result.status)
+                                status = true;
+                        break;
                     }
                 }
 
+                if(!status)
+                    foreach (var admin in admins)
+                    {
+                        if (admin.user.id != TgApi().getDefaultApiConnection().getMe().id)
+                        {
+                            var result = TgApi.getDefaultApiConnection().sendMessage(
+                                Temp.ReportGroupID,
+                                "[加群測試(不用理會此訊息)](tg://user?id=" + admin.user.id.ToString() + ")",
+                                ParseMode: TgApi.PARSEMODE_MARKDOWN);
+                            if (result.ok)
+                            {
+                                TgApi.deleteMessage(Temp.ReportGroupID, result.message_id);
+                                status = true;
+                                break;
+                            }
+                        }
+                    }
                 if (status)
                 {
                     System.Console.WriteLine("[checkHelper] Admin in report group GID : " + ChatID.ToString());
